checkpointing after untangling a day or twos worth of skittish code changes into something simpler / more direct This introduces the `ISOFileWrapper` and `InternalISOFileWrapper` subtypes of `ArtifactWrapper` which know how to parse a `ISOFile` and present it with the `ArtifactWrapper` interface Note that a `ISOFileWrapper` can only represent the "outside"- that is the actual handle on the `.iso` file. Anything *inside* of a `.iso` file is an `InternalISOFileWrapper`
Much of the existing pre-iso code assumed all files were wrapped in a `FileWrapper`:
there is now a selective apply on the `FileWrapper` object:
`FileWrapper.fromFile`, which takes a `java.io.File` and either returns
an `ISOFileWrapper` if the filename ends with `.iso` or a `FileWrapper`
for anything else. This approach is entirely dependent on file
extension, we may want to look at something like MIME Magic later for
sanity.